backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg[binary]>=3.1
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0
bcrypt>=4.1.0
requests>=2.31.0
faker>=21.0.0
openai>=1.6.0
